Changed epoll to level triggered for < 2.6.17, edge triggered for newer

+++ src/cc/Hypertable/Lib/BlockCompressionHeader.h
@@ -32,10 +32,10 @@ namespace Hypertable {
 
     static const size_t LENGTH = 26;
 
-    BlockCompressionHeader() : m_data_length(0), m_data_zlength(0), m_data_checksum(0), m_compression_type(-1)
+    BlockCompressionHeader() : m_data_length(0), m_data_zlength(0), m_data_checksum(0), m_compression_type((uint16_t)-1)
     { return; }
 
-    BlockCompressionHeader(const char *magic) : m_data_length(0), m_data_zlength(0), m_data_checksum(0), m_compression_type(-1)
+    BlockCompressionHeader(const char *magic) : m_data_length(0), m_data_zlength(0), m_data_checksum(0), m_compression_type((uint16_t)-1)
       { memcpy(m_magic, magic, 10); }
 
     virtual ~BlockCompressionHeader() { return; }
